+++ how_to_use.py
@@ -1,30 +1,25 @@
-from linear_model.LinearRegression import SimpleLinearRegression # Importing Regression
-from preprocessing.Scaler import MinMaxScaler # Importing Scaler
+from linear_model import LinearRegression # Importing Regression
+from preprocessing import MinMaxScaler # Importing Scaler
 import numpy as np
-from sklearn.linear_model import LinearRegression
 
 
 X = np.array([10 , 20, 30, 40, 50, 60])
 y = np.array([100, 200, 300 ,400 ,500, 600])
 X_test = np.array([60, 70, 80])
 
-# Without Scaler - Not Prefered
+# Without Scaler - Not Prefered Or Change Learning_rate to 0.0001 or lower value
 
-Model = SimpleLinearRegression()
-Model_sk = LinearRegression()
+Model = LinearRegression()
 
 Model.fit(X, y)
 print(Model.get_weight_bias())
 print(Model.predict(X_test))
 
-Model_sk.fit(X.reshape(-1, 1), y.reshape(-1, 1))
-print(Model_sk.coef_, Model_sk.intercept_)
-print(Model_sk.predict(X_test.reshape(-1, 1)))
 
 # ___________________________________________
 # With Scaler - Prefered
 
-Scaler = MinMaxScaler()
+Scaler = MinMaxScaler() # you can also use StandardScaler
 X = Scaler.fit_transform(X.reshape(-1, 1))
 X_test = Scaler.transform(X_test.reshape(-1, 1))