+++ model_selection/__init__.py
@@ -1,0 +1,2 @@
+from ._split import train_test_split
+__all__ = ["train_test_split"]